Conditionally define classes, not just includes.
This is required to work around an issue when Composer optimises the
autoloader into a static map: it will use the first Handler it finds,
rather than the one that's correct for the given Monolog version.

Eventually, we'll drop support for Monolog 1, and we can get rid of all
this hackery.